requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Auth
python-jose[cryptography]==3.3.0

# Config
python-dotenv==1.0.1
pydantic-settings==2.7.0

# LLM
groq>=0.9.0

# PDF extraction
pymupdf
pytesseract>=0.3.10
Pillow>=10.0.0

# RAG / Vector store
chromadb
llama-index
llama-index-vector-stores-chroma
sentence-transformers

# Hybrid search — BM25 keyword layer for exact-match retrieval
rank-bm25>=0.2.2

# Layout-aware PDF parsing — preserves tables as Markdown pipe tables
pymupdf4llm>=0.0.17
